Quadrant System

+++ Asteroids3D/Assets/Scripts/Data/QuadrantData.cs
@@ -4,5 +4,5 @@ using Unity.Mathematics;
 [GenerateAuthoringComponent]
 public struct QuadrantData : IComponentData
 {
-    public float3 quadrant;
+    public int3 quadrant;
 }
